wallet: refactor legacy logic from createtx and chainntfns
This commit continues the wallet package refactoring by moving legacy
logic from createtx.go, chainntfns.go, and import.go to deprecated.go.

chainntfns.go (legacy RPC notifications) and import.go (legacy imports)
were moved entirely to deprecated.go and deleted.

createtx.go was split: legacy logic moved to deprecated.go, while
helpers used by the new TxCreator were moved to tx_creator.go. The file
was deleted.
